Send bwid when sending events to Ophan
To Ophan, two events with the same pageviewId are considered the same event.
In the Data Lake though, "uniqueness" is given by (pageviewId, browserId).
The browserId is usually sourced by Ophan directly from the cookies, which we don't have in this case.
If the cookie value isn't found, Ophan will make up one.
This means that if we send multiple events with the same pageviewId but without a browserId, they will be
counted as one event in Ophan but multiple events in the Data Lake.

Explicitly setting a browserId should fix this.